Calculate fps as double-precision to make switching to double-precision fps values easier.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@26504 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_avs.c
@@ -299,7 +299,7 @@ static demuxer_t* demux_open_avs(demuxer_t* demuxer)
         
         //sh_video->format = get_mmioFOURCC(AVS->video_info);
         sh_video->format = mmioFOURCC('Y', 'V', '1', '2');
-        sh_video->fps = (float) ((float) AVS->video_info->fps_numerator / (float) AVS->video_info->fps_denominator);
+        sh_video->fps = (double) AVS->video_info->fps_numerator / (double) AVS->video_info->fps_denominator;
         sh_video->frametime = 1.0 / sh_video->fps;
         
         sh_video->bih = malloc(sizeof(BITMAPINFOHEADER) + (256 * 4));
